feat: Build UI for unbalanced progress bar
Created a user interface for an unbalanced progress bar. The implementation involves passing a list of values for each target, where each step exhibits uneven progress. For instance, if the steps are defined as [10, 20, 40, 70] and the current value is 30, the progress will be accurately represented between 20 and 40, rather than at 30% of 70%.